+++ src/app_runtime/app_lifecycle.h
@@ -37,6 +37,7 @@ struct AppTeardownResult {
     std::size_t cancelled_requests = 0;
     std::size_t discarded_completions = 0;
     std::size_t released_handles = 0;
+    std::size_t released_font_resources = 0;
 };
 
 struct AppCompletionPumpResult {